Moved ucla task creation to a seperate django app
