+++ corehq/apps/app_manager/forms.py
@@ -33,9 +33,3 @@ class CopyApplicationForm(forms.Form):
         if domain is None:
             raise forms.ValidationError("A valid project space is required.")
         return domain_name
-
-class UclaTaskCreationForm(forms.Form):
-    question_ids = forms.CharField(
-        widget=forms.Textarea,
-        label="Enter Question ids, one per line"
-    )
