+++ corehq/apps/app_manager/urls.py
@@ -122,6 +122,5 @@ urlpatterns = patterns('corehq.apps.app_manager.views',
         include('corehq.apps.app_manager.download_urls')),
     url(r'^formdefs/(?P<app_id>[\w-]+)/', 'formdefs', name='formdefs'),
 
-    url(r'ucla-task-creation/(?P<app_id>[\w-]+)/modules-(?P<module_id>[\w-]+)/forms-(?P<form_id>[\w-]+)/$',
-        'ucla_task_creation', name='ucla_task_creation')
+    url(r'^', include('custom.ucla.urls')),
 )
