+++ corehq/apps/app_manager/views.py
@@ -28,7 +28,7 @@ from corehq.apps.app_manager.exceptions import (
     RearrangeError,
 )
 
-from corehq.apps.app_manager.forms import CopyApplicationForm, UclaTaskCreationForm
+from corehq.apps.app_manager.forms import CopyApplicationForm
 from corehq.apps.app_manager import id_strings
 from corehq.apps.app_manager.templatetags.xforms_extras import trans
 from corehq.apps.commtrack.models import Program
@@ -2582,128 +2582,11 @@ def upload_translations(request, domain, app_id):
 
     return HttpResponseRedirect(reverse('app_languages', args=[domain, app_id]))
 
-@require_deploy_apps
-def ucla_task_creation(request, domain, app_id, module_id, form_id):
-    '''
-    This view is meant to be a run as a one-off script to support a specific
-    app that Jeremy is writing. Running this script subsequent times on the
-    same form will not have adverse affects.
-    :param request:
-    :param domain:
-    :param app_id:
-    :param module_id:
-    :param form_id:
-    :return:
-    '''
-    if request.method == 'POST':
-
-        html_form = UclaTaskCreationForm(request.POST)
-        if html_form.is_valid():
-
-            questions = html_form.cleaned_data['question_ids'].split()
-            form = get_app(domain, app_id).modules[int(module_id)].forms[int(form_id)]
-            # Make changes to the form
-            message = _ucla_form_modifier(form, questions)
-            return HttpResponse(message, content_type="text/plain")
-
-        return HttpResponse("Soemthing was wrong with the form you submitted. Your CommCare form is unchanged.", content_type="text/plain")
-
-    elif request.method == 'GET':
-        html_form = UclaTaskCreationForm()
-        response = HttpResponse()
-        response.write('<form action="'+reverse('ucla_task_creation', args=(domain, app_id, module_id, form_id))+'" method="post">')
-        response.write(html_form.as_p())
-        response.write('<p><input type="submit" value="Process Form"></p>')
-        response.write("</form>")
-        return response
-    else:
-        return HttpResponse("GET or POST only.", content_type="text/plain")
 
 common_module_validations = [
     (lambda app: app.application_version == APP_V1,
      _('Please upgrade you app to > 2.0 in order to add a Careplan module'))
 ]
-
-
-def _ucla_form_modifier(form, question_ids):
-
-    message = ""
-
-    xform = form.wrapped_xform()
-
-    # Get the questions specified in question_ids
-    question_dict = {q["value"]:FormQuestion.wrap(q) for q in form.get_questions(["en"])}
-    question_ids = {"/data/" + q for q in question_ids}.intersection(question_dict.keys())
-    questions = [question_dict[k] for k in question_ids]
-
-    # Get the existing subcases
-    existing_subcases = {c.case_name:c for c in form.actions.subcases}
-
-    for question in questions:
-        for option in question.options:
-
-            hidden_value_path = question.value + "-" + option.value
-            hidden_value_text = option.label
-
-            # Create new hidden values for each question option if they don't already exist:
-
-            if hidden_value_path not in question_dict:
-
-                # Add data element
-                tag = "{x}%s" % hidden_value_path.replace("/data/", "")
-                element = etree.Element(tag.format(**namespaces))
-                xform.data_node.append(element)
-
-                # Add bind
-                xform.itext_node.addprevious(_make_elem("bind",{
-                    "nodeset": xform.resolve_path(hidden_value_path),
-                    "calculate": '"'+hidden_value_text+'"'
-                }))
-
-                message += "Node " + hidden_value_path + " created!\n"
-            else:
-                message += "Node " + hidden_value_path + " already exists, skipping.\n"
-
-            # Create FormActions for opening subcases
-
-            if hidden_value_path not in existing_subcases:
-                action = OpenSubCaseAction(
-                    condition=FormActionCondition(
-                        type='if',
-                        question=question.value,
-                        operator='selected',
-                        answer=option.value,
-                    ),
-                    case_name=hidden_value_path,
-                    case_type='task',
-                    # Note, the case properties will not necessarily be created in the order given.
-                    case_properties={
-                        'task_responsible': '/data/task_responsible',
-                        'task_due': '/data/task_due',
-                        'owner_id': '/data/owner_id',
-                        'task_risk_factor': '/data/task_risk_factor',
-                        'study_id': '/data/study_id',
-                        'patient_name': '/data/patient_name'
-                    },
-                    close_condition=FormActionCondition(
-                        answer=None,
-                        operator=None,
-                        question=None,
-                        type='never'
-                    )
-                )
-                form.actions.subcases.append(action)
-                message += "OpenSubCaseAction " + hidden_value_path + " created!\n"
-            else:
-                message += "OpenSubCaseAction " + hidden_value_path + " already exists, skipping.\n"
-
-    app = form.get_app()
-    # Save the xform modifications
-    save_xform(app, form, etree.tostring(xform.xml, encoding="unicode"))
-    # save the action modifications
-    app.save()
-    message += "Form saved.\n"
-    return message
 
 
 FN = 'fn'
